backend/requirements.txt
# Backend Python Dependencies

# FastAPI - Modern web framework
fastapi==0.104.1

# Uvicorn - ASGI server to run FastAPI
uvicorn[standard]==0.24.0

# SQLAlchemy - Database ORM (Object-Relational Mapping)
sqlalchemy==2.0.23

# Pydantic Settings - Environment variable management
pydantic-settings==2.1.0

# Python-multipart - For form data and file uploads
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI endpoints

# Requests - For making HTTP requests (used in our test script)
requests==2.31.0
